% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlike.R
\name{remove_PACKAGES}
\alias{remove_PACKAGES}
\title{Remove package from a package database}
\usage{
remove_PACKAGES(files, dir = ".")
}
\arguments{
\item{files}{Files to remove. They must still exist at the time this
function is called.}

\item{dir}{Package directory.}
}
\description{
The files will be first removed from the database, and then
from the directory.
}
\seealso{
Other PACKAGES manipulation: \code{\link{add_PACKAGES}},
  \code{\link{update_PACKAGES}}
}
\concept{PACKAGES manipulation}
