% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{with_db}
\alias{with_db}
\title{Perform a DB query, without explicit locking}
\usage{
with_db(db_file, expr)
}
\arguments{
\item{db_file}{File of the DB.}

\item{expr}{Expression to evaluate, it can refer to the connection
handle as \code{db}.}
}
\description{
This is for read operations. They can also be called from
within a transaction. In this case the database handle will
be reused.
}
\keyword{internal}
