% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{with_db_lock}
\alias{with_db_lock}
\title{Perform a DB query, with locking}
\usage{
with_db_lock(db_file, expr)
}
\arguments{
\item{db_file}{File of the DB.}

\item{expr}{Expression to evaluate, it can refer to the connection
handle as \code{db}.}
}
\description{
This creates a transaction, and an exclusive lock.
It always creates a new DB connection, and closes it on exit.
}
\keyword{internal}
