% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{add_PACKAGES}
\alias{add_PACKAGES}
\title{Add R packages to the package database}
\usage{
add_PACKAGES(files, dir = ".")
}
\arguments{
\item{files}{Files to add, only the file names, without the path.
You can use [base::basename()] to chop off the path.}

\item{dir}{Package directory.}
}
\description{
The files must exist in the directory.
}
\seealso{
Other PACKAGES manipulation: \code{\link{remove_PACKAGES}},
  \code{\link{update_PACKAGES}}
}

