% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranlike.R
\name{create_empty_PACKAGES}
\alias{create_empty_PACKAGES}
\title{Create an empty package database}
\usage{
create_empty_PACKAGES(dir = ".", fields = NULL)
}
\arguments{
\item{dir}{Character vector describing the location of the repository
    (directory including source or binary packages) to generate the
    \file{PACKAGES}, \file{PACKAGES.gz} and \file{PACKAGES.rds} files
    from and write them to.
  }

\item{fields}{
    a character vector giving the fields to be used in the
    \file{PACKAGES}, \file{PACKAGES.gz} and \file{PACKAGES.rds} files in
    addition to the default ones, or \code{NULL} (default).

    The default corresponds to
    the fields needed by \code{\link{available.packages}}:
    \code{"Package"}, \code{"Version"}, \code{"Priority"},
    \code{"Depends"}, \code{"Imports"}, \code{"LinkingTo"},
    \code{"Suggests"}, \code{"Enhances"}, \code{"OS_type"},
    \code{"License"} and \code{"Archs"}, and those fields will always
    be included, plus
    the file name in field \code{"File"} if \code{addFile = TRUE}
    and the path to the subdirectory in field \code{"Path"} if subdirectories
    are used.
  }
}
\description{
Create an empty package database if it does not exist.
It also updates the PACKAGES* files from the new (empty) database.
}
